port transaction signing test to jellyfish-txsignature
