+++ packages/jellyfish-transaction/src/index.ts
@@ -12,4 +12,3 @@ export * from './script'
 export * from './tx'
 export * from './tx_composer'
 export * from './tx_segwit'
-export * from './tx_signature'
